LayoutTests/fast/mediastream/RTCPeerConnection-createOffer-expected.txt

Tests RTCPeerConnection createOffer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.createOffer(requestSucceeded1, requestFailed1, {mandatory:{"succeed":true}}); did not throw exception.
PASS createOffer request succeeded.
PASS sessionDescription.type is "offer"
PASS pc.createOffer(requestSucceeded2, requestFailed2, {mandatory:{"succeed":false}}); did not throw exception.
PASS createOffer request failed.
PASS errorReason.name is "IncompatibleConstraintsError"
PASS pc.createOffer(requestSucceeded3, requestFailed1); did not throw exception.
PASS createOffer request succeeded.
PASS sessionDescription.type is "offer"
PASS successfullyParsed is true

TEST COMPLETE
